films/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
boto3==1.26.103
botocore==1.29.103
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.1.7
django-filter==23.2
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
exceptiongroup==1.1.1
executing==1.2.0
greenlet==2.0.2
idna==3.4
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
msgpack==1.0.5
mypy==1.2.0
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
pynvim==0.4.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-cache==1.0.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
url-normalize==1.4.3
urllib3==1.26.15
wcwidth==0.2.6
